+++ src/main/java/com/gitlab/cdagaming/craftpresence/ModLogger.java
@@ -1,5 +1,6 @@
 package com.gitlab.cdagaming.craftpresence;
 
+import com.gitlab.cdagaming.craftpresence.handler.StringHandler;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -13,14 +14,26 @@ public class ModLogger {
     }
 
     public void error(final String logMessage, Object... logArguments) {
-        logInstance.error(logMessage, logArguments);
+        if (CraftPresence.player != null && !CraftPresence.CONFIG.hasChanged && CraftPresence.CONFIG.showLoggingInChat) {
+            StringHandler.sendMessageToPlayer(CraftPresence.player, "§6§l[§f" + loggerName + "§6]§r " + logMessage);
+        } else {
+            logInstance.error(logMessage, logArguments);
+        }
     }
 
     public void warn(final String logMessage, Object... logArguments) {
-        logInstance.warn(logMessage, logArguments);
+        if (CraftPresence.player != null && !CraftPresence.CONFIG.hasChanged && CraftPresence.CONFIG.showLoggingInChat) {
+            StringHandler.sendMessageToPlayer(CraftPresence.player, "§6§l[§f" + loggerName + "§6]§r " + logMessage);
+        } else {
+            logInstance.warn(logMessage, logArguments);
+        }
     }
 
     public void info(final String logMessage, Object... logArguments) {
-        logInstance.info(logMessage, logArguments);
+        if (CraftPresence.player != null && !CraftPresence.CONFIG.hasChanged && CraftPresence.CONFIG.showLoggingInChat) {
+            StringHandler.sendMessageToPlayer(CraftPresence.player, "§6§l[§f" + loggerName + "§6]§r " + logMessage);
+        } else {
+            logInstance.info(logMessage, logArguments);
+        }
     }
 }